refactor utils pkg
